don't edit geo-types. remove another box. add comment sections.
