+++ geo/src/algorithm/coords_iter.rs
@@ -10,6 +10,10 @@ pub trait CoordsIter<'a, T: CoordinateType + 'a> {
     fn coords_iter(&'a self) -> Self::Iter;
 }
 
+// ┌──────────────────────────┐
+// │ Implementation for Point │
+// └──────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Point<T> {
     type Iter = iter::Once<Coordinate<T>>;
 
@@ -17,6 +21,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Point<T> {
         iter::once(self.0)
     }
 }
+
+// ┌─────────────────────────┐
+// │ Implementation for Line │
+// └─────────────────────────┘
 
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Line<T> {
     type Iter = iter::Chain<iter::Once<Coordinate<T>>, iter::Once<Coordinate<T>>>;
@@ -26,6 +34,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Line<T> {
     }
 }
 
+// ┌───────────────────────────────┐
+// │ Implementation for LineString │
+// └───────────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for LineString<T> {
     type Iter = iter::Copied<slice::Iter<'a, Coordinate<T>>>;
 
@@ -34,26 +46,78 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for LineString<T> {
     }
 }
 
+// ┌────────────────────────────┐
+// │ Implementation for Polygon │
+// └────────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Polygon<T> {
     type Iter = iter::Chain<
         <geo_types::LineString<T> as CoordsIter<'a, T>>::Iter,
-        geo_types::InteriorCoordsIter<'a, T>,
+        InteriorCoordsIter<'a, T>,
     >;
 
     fn coords_iter(&'a self) -> Self::Iter {
         self.exterior()
             .coords_iter()
-            .chain(self.interior_coords_iter())
+            .chain(
+                InteriorCoordsIter(
+                    LineStringsToCoordsIter(self.interiors().iter()).flatten()
+                )
+            )
     }
 }
+
+// Utility to transform Iterator<LineString> into Iterator<&[Coordinate]>
+struct LineStringsToCoordsIter<'a, T: CoordinateType>(slice::Iter<'a, LineString<T>>);
+
+impl<'a, T: CoordinateType> Iterator for LineStringsToCoordsIter<'a, T> {
+    type Item = &'a [Coordinate<T>];
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next().map(|n| &n.0[..])
+    }
+}
+
+
+// Utility to transform Iterator<&[Coordinate]> into Iterator<Coordinate>
+#[doc(hidden)]
+pub struct InteriorCoordsIter<'a, T: CoordinateType>(iter::Flatten<LineStringsToCoordsIter<'a, T>>);
+
+impl<'a, T: CoordinateType> Iterator for InteriorCoordsIter<'a, T> {
+    type Item = Coordinate<T>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next().copied()
+    }
+}
+
+// ┌───────────────────────────────┐
+// │ Implementation for MultiPoint │
+// └───────────────────────────────┘
 
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiPoint<T> {
-    type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
+    type Iter = MultiPointCoordsIter<'a, T>;
 
     fn coords_iter(&'a self) -> Self::Iter {
-        Box::new(self.0.iter().flat_map(|m| m.coords_iter()))
+        MultiPointCoordsIter(self.0.iter())
     }
 }
+
+// Utility to transform Iterator<Point> into Iterator<Coordinate>
+#[doc(hidden)]
+pub struct MultiPointCoordsIter<'a, T: CoordinateType>(slice::Iter<'a, Point<T>>);
+
+impl<'a, T: CoordinateType> Iterator for MultiPointCoordsIter<'a, T> {
+    type Item = Coordinate<T>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.0.next().map(|point| point.0)
+    }
+}
+
+// ┌────────────────────────────────────┐
+// │ Implementation for MultiLineString │
+// └────────────────────────────────────┘
 
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiLineString<T> {
     type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
@@ -63,6 +127,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiLineString<T> {
     }
 }
 
+// ┌─────────────────────────────────┐
+// │ Implementation for MultiPolygon │
+// └─────────────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiPolygon<T> {
     type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
 
@@ -71,6 +139,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiPolygon<T> {
     }
 }
 
+// ┌───────────────────────────────────────┐
+// │ Implementation for GeometryCollection │
+// └───────────────────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for GeometryCollection<T> {
     type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
 
@@ -78,6 +150,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for GeometryCollection<T> {
         Box::new(self.0.iter().flat_map(|m| m.coords_iter()))
     }
 }
+
+// ┌─────────────────────────┐
+// │ Implementation for Rect │
+// └─────────────────────────┘
 
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Rect<T> {
     type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
@@ -104,6 +180,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Rect<T> {
     }
 }
 
+// ┌─────────────────────────────┐
+// │ Implementation for Triangle │
+// └─────────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Triangle<T> {
     type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
 
@@ -116,6 +196,10 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Triangle<T> {
     }
 }
 
+// ┌─────────────────────────────┐
+// │ Implementation for Geometry │
+// └─────────────────────────────┘
+
 impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Geometry<T> {
     type Iter = Box<dyn Iterator<Item = Coordinate<T>> + 'a>;
 
@@ -125,7 +209,7 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Geometry<T> {
             Geometry::Line(g) => Box::new(g.coords_iter()),
             Geometry::LineString(g) => Box::new(g.coords_iter()),
             Geometry::Polygon(g) => Box::new(g.coords_iter()),
-            Geometry::MultiPoint(g) => g.coords_iter(),
+            Geometry::MultiPoint(g) => Box::new(g.coords_iter()),
             Geometry::MultiLineString(g) => g.coords_iter(),
             Geometry::MultiPolygon(g) => g.coords_iter(),
             Geometry::GeometryCollection(g) => g.coords_iter(),
